+++ HISTORY.md
@@ -25,7 +25,7 @@
 * **Notable Changes:**
     * significantly sped-up rendering morphs representing data inside table cells
     * support testing selectors (translatable text) for equality with text and numbers without needing to "unselect" them
-    * added support for color slot default values in primitives
+    * added support for color slot default values
 * **Notable Fixes:**
     * fixed sorting the distribution of compound non-list data (blocks, colors)
 * **Translation Updates:**
@@ -33,6 +33,7 @@
 
 ### 2025-04-15
 * blocks, objects: added support for color slot default values in primitives
+* blocks, byob: added support for color slot default values in custom blocks
 
 ### 2025-04-14
 * updated German translation for "darker" and "lighter"